Contributing upstream

+++ builders/repo_config.sh
@@ -5,14 +5,20 @@
 set -e
 set -x
 
+. $( dirname $0 )/common
+
 tmpdir=$( mktemp -d )
 trap "echo removing ${tmpdir}; rm -rf ${tmpdir}" EXIT
 
-repofile="/etc/yum.repos.d/${BUCKET}-${REPO}.repo"
+PKG_NAME="${BUCKET}"
+PKG_VER="1.0"
 
-mkdir -p $( dirname ${tmpdir}/${repofile} )
+if ! pkg_exists_in_repo ${PKG_NAME}-${PKG_VER}; then
+    repofile="/etc/yum.repos.d/${BUCKET}-${REPO}.repo"
 
-cat << EOF > ${tmpdir}/${repofile}
+    mkdir -p $( dirname ${tmpdir}/${repofile} )
+
+    cat << EOF > ${tmpdir}/${repofile}
 [${BUCKET}-${REPO}]
 name=My S3 repo
 baseurl=https://${BUCKET}.s3.amazonaws.com/${REPO}/
@@ -20,12 +26,13 @@ enabled=1
 gpgcheck=0
 EOF
 
-fpm \
-    -s dir \
-    -t rpm \
-    -n ${BUCKET} \
-    -v 1.0 \
-    -a noarch \
-    --description "Yum repo config for s3://${BUCKET}/${REPO}" \
-    -C ${tmpdir} \
-    ${repofile#/}
+    fpm \
+        -s dir \
+        -t rpm \
+        -n ${PKG_NAME} \
+        -v ${PKG_VER} \
+        -a noarch \
+        --description "Yum repo config for s3://${BUCKET}/${REPO}" \
+        -C ${tmpdir} \
+        ${repofile#/}
+fi
